Use cast for `off_t` since BSD uses a different type (`long long`) than GNU (`long int`)

+++ cache/compression.c
@@ -240,8 +240,8 @@ int tryLoad(const char *fileName, struct FCVersion *version, const char *ext,
 	version->data = malloc(status.st_size);
 	if (version->data == NULL) {
 		fprintf(stderr, ANSI_COLOR_RED"[Cache::Compression::tryLoad] Failed to"
-				" allocate %li octets for file '%s' on extension '%s'"
-				ANSI_COLOR_RESETLN, status.st_size, buf, ext);
+				" allocate %lli octets for file '%s' on extension '%s'"
+				ANSI_COLOR_RESETLN, (long long int) status.st_size, buf, ext);
 		close(fd);
 		return 0;
 	}
